Add priority icons and do more refactoring.
Now priority icons are shown before type icons. In addition the issue title is not colored anymore.
The project is not stored in the localStorage, the ID is used as location hash instead. This make it possible to open multiple projects in multiple tabs.
The rendering was refactored and a lot of duplicated code was removed.